+++ content.php
@@ -9,13 +9,13 @@
 ?>
 <article id="post-<?php the_ID(); ?>" <?php post_class(); ?> role="article">
     <header>
-    <?php if ( is_single() ) : ?>
-    <h1 class="entry-title"><?php the_title(); ?></h1>
-    <?php else : ?>
-    <h2 class="entry-title">
-    <a href="<?php the_permalink(); ?>" title="<?php echo esc_attr( sprintf( __( 'View %s', 'narga' ), the_title_attribute( 'echo=0' ) ) ); ?>" rel="bookmark"><?php the_title(); ?></a></h2>
-    <?php endif; // is_single() ?>
-    <?php narga_entry_meta(); ?>
+        <?php if ( is_single() ) : ?>
+        <h1 class="entry-title"><?php the_title(); ?></h1>
+        <?php else : ?>
+        <h2 class="entry-title">
+            <a href="<?php the_permalink(); ?>" title="<?php echo esc_attr( sprintf( __( 'View %s', 'narga' ), the_title_attribute( 'echo=0' ) ) ); ?>" rel="bookmark"><?php the_title(); ?></a></h2>
+        <?php endif; // is_single() ?>
+        <?php narga_entry_meta(); ?>
     </header>
     <section class="entry-content<?php if ( has_post_format( 'video' )) {echo ' flex-video';}?>">
         <?php the_content( __( 'Continue reading <span class="meta-nav">&rarr;</span>', 'narga' ) ); ?>