+++ page.php
@@ -14,9 +14,14 @@
 <?php get_header(); ?>
 <!-- Row for main content area -->
 <div id="page-content-wrapper" class="large-8 columns" role="content">
-        <?php narga_breadcrumb(); ?>
-        <?php get_template_part('content', 'page'); ?>
-        <?php comments_template( '', true ); ?>
-</div><!-- End Content row -->
+<?php #Breadcrumb Control
+if (narga_options('breadcrumb') == 1) :
+    narga_breadcrumb();
+endif;
+?>
+<?php get_template_part('content', 'page'); ?>
+<?php comments_template( '', true ); ?>
+</div>
+<!-- End Content row -->
 <?php get_sidebar(); ?>
 <?php get_footer(); ?>